requirements.txt
requests~=2.32
pyrogram==2.0.106
TgCrypto==1.2.5
pydantic==2.11.4
PyYAML~=6.0.2
uvicorn==0.34.2
fastapi==0.115.12
aiohttp~=3.11.18
starlette~=0.46
prometheus_client~=0.21
